Removed big decimals from backend

+++ command-processor/src/main/scala/compman/compsrv/logic/schedule/StageGraph.scala
@@ -48,7 +48,7 @@ case class StageGraph private[schedule](
 
   def getStageId(id: String): String = { fightsMap(id).getStageId }
 
-  def getDuration(id: String): BigDecimal = { fightsMap(id).getDuration }
+  def getDuration(id: String): Int = { fightsMap(id).getDuration }
 
   def getStageFights(stageId: String): List[String] = {
     if (stageIdsToIds.containsKey(stageId)) { stageIdsToFightIds(stageIdsToIds.get(stageId)) }
