+++ competition-serv-model/src/main/java/compman/compsrv/model/dto/competition/PointGroupDTO.java
@@ -4,7 +4,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
 
-import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -12,6 +11,6 @@ import java.math.BigDecimal;
 public class PointGroupDTO {
     private String id;
     private String name;
-    private BigDecimal priority;
-    private BigDecimal value;
+    private Integer priority;
+    private Integer value;
 }
