+++ competition-serv-model/src/main/java/compman/compsrv/model/dto/competition/PromoCodeDTO.java
@@ -4,7 +4,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 
 @Data
@@ -12,7 +11,7 @@ import java.time.Instant;
 @Accessors(chain = true)
 public class PromoCodeDTO {
     private String id;
-    private BigDecimal coefficient;
+    private Integer coefficient;
     private String competitionId;
     private Instant startAt;
     private Instant expireAt;
